+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Changelog
 * Add ``mode`` argument to :meth:`api.utils.list_variants` method.
 * Update :command:`create-consolidated-vcf` command to implement phase-extension algorithm.
 * Remove ``SO`` and ``Type`` columns from the variant table.
+* Update :class:`api.genotype.GSTM1Genotyper` class.
 * Update variant information for following alleles: CYP2B6*17, CYP2D6*15, CYP2D6*21.
 
 0.4.1 (2021-09-21)